Update module structure
