+++ src/lib.rs
@@ -5,7 +5,7 @@
 
 // #![deny(missing_docs)]
 
-/// Compositor component.
+/// Verso's compositor component to handle webrender.
 pub mod compositor;
 /// Utilities to read options and preferences.
 pub mod config;
@@ -13,6 +13,8 @@ pub mod config;
 pub mod errors;
 /// Utilities to handle keyboard inputs and states.
 pub mod keyboard;
+/// Utilities to handle keyboard inputs and states.
+pub mod touch;
 /// Main entry types and functions.
 pub mod verso;
 /// Web view types to handle web browsing contexts.
